pyproject.toml
[project]
name = "CNN_CIFAR"
version = "0.1.0"
description = "A Django-based image classification web application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crispy-tailwind>=1.0.3",
    "django>=5.1.7",
    "django-ratelimit>=4.1.0",
    "pillow>=11.1.0",
    "python-decouple>=3.8",
    "reportlab>=4.3.1",
    "requests>=2.32.3",
    "social-auth-app-django>=5.4.3",
    "tensorflow==2.18.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.6",
]
test = [
    "pytest>=8.3.3",
    "pytest-django>=4.9.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["."]
exclude = [
    ".venv",
    "tests",
    "__pycache__",
    "build",
    "dist",
    ".git",
    "migrations",
    "*/migrations/*",
    "media",
    "static",
    "node_modules",
    "*.sqlite3",
    "cnn_tf.ipynb"
]
preview = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # bugbear
    "C",    # complexity
    "N",    # naming conventions
    "UP",   # pyupgrade
    "A",    # flake8-builtins
    "C4",   # comprehensions
    "RUF",  # Ruff-specific rules
    "DJ",   # flake8-django (Django-specific linting)
    "Q",    # flake8-quotes (consistent quote usage)
    "S",    # flake8-bandit (security issues)
    "T10",  # flake8-debugger (detect debugger imports)
    "T20",  # flake8-print (detect print statements)
    "PLC",  # pylint convention rules
    "PLE",  # pylint error rules
    "PLR",  # pylint refactoring suggestions
]
ignore = [
    "E501",    # Line length (handled by line-length setting)
    "D105",    # Missing docstring in magic method
    "D107",    # Missing docstring in __init__
    "D203",    # 1 blank line required before class docstring (conflicts with D211)
    "D213",    # Multi-line docstring summary should start at the second line (conflicts with D212)
    "S105",    # Possible hardcoded password (false positives in Django settings)
    "DJ001",   # Avoid using null=True on string fields (common in Django models)
    "PLR2004", # Magic value used in comparison (can be too strict for some Django logic)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "C901", "S101"]
"scripts/*" = ["D", "E501"]
"*/migrations/*" = ["D", "E501", "N", "F", "C901"]
"manage.py" = ["D", "E501", "N"]
"*/settings.py" = ["D", "S105", "N"]
"*/urls.py" = ["N"]

[tool.ruff.lint.isort]
known-first-party = ["account", "prediction", "imgpredict"]
known-third-party = ["django", "crispy_tailwind", "pillow", "reportlab"]
section-order = ["future", "standard-library", "third-party", "first-party", "local"]
lines-after-imports = 2
